pyproject.toml
[tool.poetry]
name = "intersection"
version = "0.1.0"
description = "Telegram based version of a text based game where two players try to find the same word."
authors = ["gruvw <gruvw.dev@gmail.com>"]
readme = "README.md"
homepage = "https://github.com/gruvw/intersection"
packages = [
    { include = "intersection" }
]

[tool.poetry.scripts]
intersection = 'intersection.cli:start'

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^13.13"
click = "^8.1.3"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
